Case_02/api/requirements.txt
Package             Version
------------------- --------
asgiref             3.7.2
certifi             2023.5.7
charset-normalizer  3.1.0
coreapi             2.3.3
coreschema          0.0.4
Django              4.2.2
djangorestframework 3.14.0
drf-yasg            1.21.5
idna                3.4
inflection          0.5.1
itypes              1.2.0
Jinja2              3.1.2
MarkupSafe          2.1.3
packaging           23.1
pip                 23.1.2
pytz                2023.3
requests            2.31.0
ruamel.yaml         0.17.31
ruamel.yaml.clib    0.2.7
setuptools          67.7.2
six                 1.16.0
sqlparse            0.4.4
typing_extensions   4.6.3
tzdata              2023.3
uritemplate         4.1.1
urllib3             2.0.3
wheel               0.40.0
